A frustrating exercise has led me in a circle
C# has updated their library to v1.26, but has not released the support
packages to NuGet. This is causing a build failure on master, so the
test script currently resets the C# repo to v1.25. Once the support
package is published, it should be switched back. We should also use the
new BuildGenerated.sh script they've included.