Add the git cookie refresh daemon
This script periodically refreshes the
"$HOME/.git-credential-cache/cookie" file with an OAuth2 token obtained
from the Compute service for this app's service account.

Authentication by this method is required for direct pushes to the Go
googlesource repository, which is configured to accept unreviewed
changes to master from this app's service account.