Copy a snapshot of toolkit
Preserve the implementation of Discovery samplegen.

The contents of github.com/googleapis/toolkit/ as of 2017-10-20 12:40
are captured in /toolkit/. This is a one-time sync to cut development
costs on the Toolkit repo, which will continue to evolve. We expect few
or no changes to the Discovery samplegen code, which will henceforth
only live in this repo, not in Toolkit.