Make toolkit compile
Commit "Copy a snapshot of toolkit" 958f34e does not seem to be an exact copy of toolkit as of 2017-10-20 12:40. The toolkit in this repo does not compile.

This commit is created by running git reset --hard e4541cdfd24cbe3d6e0bc763aa7d7d3c32f010d1 and git clean -d -f. The commit hash points to the commit "Ruby: Disambiguate mock credentials classes in tests. (#1644)" in toolkit, which is right before the commit where discovery samplegen is deleted.

This commit will revert "change nodejs import to const {google}", which will be fixed in a coming PR.