fix go generation
The way this was managing the dependencies used for Go was not
honoring modules. This caused the apiaries to be built with HEAD
of deps. There was a recent change in one of our deps that broke on
us when using head. We should now use modules for building the code.
Note, the ennvar is explicitly not set for the `go get` command.